Rename formattable module
